All cureent config distributed; next step, loading
Every section of myinit.org has been moved to an init-* file with the
appropriate (provide ...) at the end so that (require ...) can be used
in init.el.
Next step is to load each file appropriately in init.el and modify the
usepackage declarations to invoke some packages - such as those for
language modes (elpy) when they are called.
Also TODO: figure out how to use the custom-file